@@ -1,6 +1,6 @@
 package com.example.toolkit
 
-import org.junit.Test
+import org.junit.jupiter.api.Test
 
 import org.junit.Assert.*
 
